python/seo-pipeline/requirements.txt
# Core
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.32.0

# Google APIs (reused from claude-seo)
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-analytics-data>=0.18.0

# HTML parsing (used by some claude-seo scripts)
beautifulsoup4>=4.12.0
lxml>=6.0.0
